Extract logic to retrieve accessibility settings.
